refactor: DDD 적용을 위한 폴더 구조 리팩토링 - 계층 간 관심사 분리 강화
