+++ hub/src/main/java/com/sparta/hub/presentation/controller/HubController.java
@@ -1,11 +1,11 @@
 package com.sparta.hub.presentation.controller;
 
-import com.sparta.hub.application.common.response.ResponseBody;
-import com.sparta.hub.application.common.response.ResponseUtil;
-import com.sparta.hub.application.dto.HubRequestDTO;
-import com.sparta.hub.application.dto.HubResponseDTO;
+import com.sparta.hub.presentation.dto.response.common.ResponseBody;
+import com.sparta.hub.presentation.dto.response.common.ResponseUtil;
+import com.sparta.hub.presentation.dto.response.HubResponseDto;
 import com.sparta.hub.domain.model.Hub;
 import com.sparta.hub.application.service.HubService;
+import com.sparta.hub.presentation.dto.request.HubRequest;
 import jakarta.validation.Valid;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
@@ -17,7 +17,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -31,31 +30,31 @@ public class HubController {
     private final HubService hubService;
 
     @PostMapping
-    public ResponseEntity<ResponseBody<HubResponseDTO>> createHub(
-            @RequestBody @Valid HubRequestDTO hubRequestDTO) {
-        Hub createdHub = hubService.createHub(hubRequestDTO.getName(), hubRequestDTO.getAddress());
-        return ResponseEntity.ok(ResponseUtil.createSuccessResponse(HubResponseDTO.fromEntity(createdHub)));
+    public ResponseEntity<ResponseBody<HubResponseDto>> createHub(
+            @RequestBody @Valid HubRequest hubRequest) {
+        Hub createdHub = hubService.createHub(hubRequest.toDTO());
+        return ResponseEntity.ok(ResponseUtil.createSuccessResponse(HubResponseDto.fromEntity(createdHub)));
     }
 
     @GetMapping("/{hubId}")
-    public ResponseEntity<ResponseBody<HubResponseDTO>> getHubById(@PathVariable UUID hubId) {
+    public ResponseEntity<ResponseBody<HubResponseDto>> getHubById(@PathVariable UUID hubId) {
         Hub hub = hubService.getHubById(hubId);
-        return ResponseEntity.ok(ResponseUtil.createSuccessResponse(HubResponseDTO.fromEntity(hub)));
+        return ResponseEntity.ok(ResponseUtil.createSuccessResponse(HubResponseDto.fromEntity(hub)));
     }
 
     @GetMapping
-    public ResponseEntity<ResponseBody<Page<HubResponseDTO>>> getAllHubs(Pageable pageable) {
+    public ResponseEntity<ResponseBody<Page<HubResponseDto>>> getAllHubs(Pageable pageable) {
         Page<Hub> hubs = hubService.getAllHubs(pageable);
-        Page<HubResponseDTO> responseDTOs = hubs.map(HubResponseDTO::fromEntity);
+        Page<HubResponseDto> responseDTOs = hubs.map(HubResponseDto::fromEntity);
         return ResponseEntity.ok(ResponseUtil.createSuccessResponse(responseDTOs));
     }
 
     @PatchMapping("/{hubId}")
-    public ResponseEntity<ResponseBody<HubResponseDTO>> updateHub(
+    public ResponseEntity<ResponseBody<HubResponseDto>> updateHub(
             @PathVariable UUID hubId,
-            @RequestBody @Valid HubRequestDTO hubRequestDTO) {
-        Hub updatedHub = hubService.updateHub(hubId, hubRequestDTO.getName(), hubRequestDTO.getAddress());
-        return ResponseEntity.ok(ResponseUtil.createSuccessResponse(HubResponseDTO.fromEntity(updatedHub)));
+            @RequestBody @Valid HubRequest hubRequest) {
+        Hub updatedHub = hubService.updateHub(hubId, hubRequest.toDTO());
+        return ResponseEntity.ok(ResponseUtil.createSuccessResponse(HubResponseDto.fromEntity(updatedHub)));
     }
 
     @DeleteMapping("/{hubId}")
@@ -67,11 +66,11 @@ public class HubController {
     }
 
     @GetMapping("/search")
-    public ResponseEntity<ResponseBody<Page<HubResponseDTO>>> searchHubsByName(
+    public ResponseEntity<ResponseBody<Page<HubResponseDto>>> searchHubsByName(
             @RequestParam String name,
             Pageable pageable) {
         Page<Hub> hubs = hubService.searchHubsByName(name, pageable);
-        Page<HubResponseDTO> responseDTOs = hubs.map(HubResponseDTO::fromEntity);
+        Page<HubResponseDto> responseDTOs = hubs.map(HubResponseDto::fromEntity);
         return ResponseEntity.ok(ResponseUtil.createSuccessResponse(responseDTOs));
     }
 }
